+++ src/renderer/impl/render_command_composer/rcc_multi_draw_elements_indirect.cpp
@@ -46,7 +46,7 @@ public:
         }
     }
 
-    bool update(uint64_t /*timestamp*/) override {
+    bool update(uint32_t /*tick*/) override {
         return false;
     }
 
@@ -72,7 +72,7 @@ public:
         }
     }
 
-    bool update(uint64_t /*timestamp*/) override {
+    bool update(uint32_t /*tick*/) override {
         return false;
     }
